requirements.txt
tensorflow
flask
gunicorn
python-chess
numpy
